refactor: move make tree to global util
